minor changes to sccoda and tasccoda
