+++ docs/usage/usage.md
@@ -155,7 +155,7 @@ See [Differential abundance testing on single-cell data using k-nearest neighbor
 
 See [milopy tutorial](https://pertpy.readthedocs.io/en/latest/tutorials/notebooks/milopy.html) for a more elaborate tutorial.
 
-### scCODA and tascCODA
+### Compositional analysis
 
 Reimplementation of scCODA for identification of compositional changes in high-throughput sequencing count data and tascCODA for sparse, tree-aggregated modeling of high-throughput sequencing data.
 See [scCODA is a Bayesian model for compositional single-cell data analysis](https://www.nature.com/articles/s41467-021-27150-6) for statistical methodology and benchmarking performance of scCODA and [tascCODA: Bayesian Tree-Aggregated Analysis of Compositional Amplicon and Single-Cell Data](https://www.frontiersin.org/articles/10.3389/fgene.2021.766405/full) for statistical methodology and benchmarking performance of tascCODA.
@@ -168,8 +168,8 @@ See [scCODA is a Bayesian model for compositional single-cell data analysis](htt
 .. autosummary::
     :toctree: tools
 
-    tools.SccodaModel2
-    tools.TasccodaModel2
+    tools.Sccoda
+    tools.Tasccoda
 ```
 
 ### Representation
@@ -221,7 +221,7 @@ See [scCODA is a Bayesian model for compositional single-cell data analysis](htt
     plot.milo.nhood_counts_by_cond
 ```
 
-### scCODA and tascCODA
+### Compositional analysis
 
 ```{eval-rst}
 .. autosummary::
