+++ pertpy/plot/__init__.py
@@ -1,4 +1,4 @@
 from pertpy.plot._augurpy import AugurpyPlot as ag  # noqa: N813
+from pertpy.plot._coda import CodaPlot as coda  # noqa: N813
 from pertpy.plot._milopy import MilopyPlot as milo  # noqa: N813
 from pertpy.plot._mixscape import MixscapePlot as ms  # noqa: N813
-from pertpy.plot._sccoda import SccodaPlot as coda  # noqa: N813
