+++ pertpy/tools/__init__.py
@@ -2,4 +2,5 @@ from pertpy.tools._augurpy import Augurpy
 from pertpy.tools._kernel_pca import kernel_pca
 from pertpy.tools._milopy import Milopy
 from pertpy.tools._mixscape import Mixscape
-from pertpy.tools._sccoda import SccodaModel2, TasccodaModel2
+from pertpy.tools._sccoda import Sccoda
+from pertpy.tools._tasccoda import Tasccoda
